refactor(parking-lot): 공공데이터에 default 이미지 추가 및  패키지 이동 #59
